Move readonly class test to PHP 8.2
