+++ tests/Fixture/PHP82/ReadOnlyClass.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PHP83;
+namespace PHP82;
 
 readonly class ReadonlyClass {
     public function foo(): string
