+++ tests/Unit/PHP83/Php83LanguageFeaturesTest.php
@@ -18,7 +18,6 @@ use PHP83\Classes;
 use PHP83\ClassName;
 use PHP83\Enums;
 use PHP83\Interfaces;
-use PHP83\ReadonlyClass;
 use PHP83\Traits;
 use Throwable;
 
@@ -89,16 +88,5 @@ final class Php83LanguageFeaturesTest extends MockeryTestCase
         $this->expectException(Exception::class);
 
         mock(Enums::class);
-    }
-
-    /**
-     * @throws Throwable
-     */
-    public function testCanNotMockReadonlyClasses(): void
-    {
-        $this->expectException(Exception::class);
-        $this->expectExceptionMessage('The class \PHP83\ReadonlyClass is marked readonly');
-
-        mock(ReadonlyClass::class);
     }
 }
